+++ Sensor/src/homography/homography.h
@@ -8,7 +8,7 @@
 #define HOMOGRAPHY_H
 
 #include "Arduino.h"
-#include "MatrixMath.h"
+#include "../MatrixMath/MatrixMath.h"
 #include "math.h"
 
     class homography
@@ -16,9 +16,9 @@
         public:
             homography();
             void calculateHomographyMatrix();
-            void calculateCoordinates(int, int);
-            uint16_t getX();
-            uint16_t getY();
+            void calculateCoordinates(double, double);
+            double getX();
+            double getY();
             void setCalArray(int,int,int);
             int getCalArray(int, int);
             void orderCalArray();
@@ -26,9 +26,9 @@
         private:
             uint8_t findMax(uint16_t ARRAY[]);
             mtx_type _H[3][3];
-            uint16_t _x;
-            uint16_t _y;
+            double _x;
+            double _y;
             uint16_t _bounds[2][4];
-            uint16_t _calArray[2][4];
+            int16_t _calArray[2][4];
     };
 #endif
